add patch-package to top level
Fragile and kind of dumb, but turns out this is necessary when running `npm run android` because the package.json in the nodejs-project directory has the postinstall step that calls this when nodejs-mobile internally calls `npm rebuild`. That module is not moved into that directory, so the rebuild will error if it can't find patch-package elsewhere in our project. Basically, including the dep in our top-level allows the postinstall to no-op.