fix: fix freeze after hot reload / fast refresh
Fixes #113

After reload, the frontend is sending a `get-server-status` message,
which the nodejs process receives, and tries to send back the current
status. However this message back never reaches the frontend, likely due
to the way the RN bridge is re-created during a reload.

One fix that worked was to add a small delay when responding to
`get-server-status`, but analysing the logs showed that, from the nodejs
process perspective, the app is going into the background and back into
the foreground after the reload, and _after_ the `get-server-status`
message is received. The fix that this implements is to re-send the
server status whenever the app comes back into the foreground, this
seems to fix things because it seems like nodejs bridge messages sent
while the app is in the background do not get through.