chore: "fix" lint task
`npm run lint` had two problems:

1. An error: "prettier.resolveConfig.sync is not a function".
2. Linting too many files

This fixes that by installing the latest version of
`eslint-plugin-prettier` as a direct dependency and by using our
existing `.gitignore` to exclude files.

There are 92 errors and 353 warnings to fix, but now we can fix them.
(Most of them should be easy to address.)