chore: improve wifi state management
We need the device's wifi state at a few points in the app. The way we
previously did this had a few disadvantages:

1. If you were connected to wifi and another network (like cellular),
   we might get the wrong network's IP address. You might also be unable
   to fetch the SSID.

2. It used two different dependencies, because neither does exactly what
   we want.

3. These dependencies rely on deprecated APIs.

This change fixes those by implementing a native module, `WifiModule`,
to address these problems. It also adds an React hook for using it.

I think this is a useful change on its own, but will also help us as we
try to [improve local peer discovery][0] in upcoming work.

[0]: https://github.com/digidem/mapeo-core-next/issues/474