chore: fix invalid invite type (#277)
This is a types-only change and should have no impact on functionality.

There was a type error when inviting someone to a project (see [these
lines][0]). This was because the `role` type was too broad; you can't
invite someone with the "LEFT" role, for example.

This fixes that by using a narrower type.

In the long term, `@mapeo/core` should probably expose these types in a
better way. In the short term, I think this is the right solution to
fix the type error.

[0]: https://github.com/digidem/comapeo-mobile/blob/92fc7f42dd42ab45b267cc0df389f447b97a7880/src/frontend/screens/Settings/ProjectSettings/YourTeam/ReviewAndInvite/index.tsx#L35-L36