chore: fix type error in ObservationsList component (#280)
This is a types-only change and should have no impact on functionality.

TypeScript complained about the type we used for `<FlatList>`'s
`getItemLayout` prop. Specifically, it complained that the first
argument was invalid.

We don't even use the first argument, so we can just use the `unknown`
type to sidestep this issue.