chore: send app diagnostic metrics
This sends app diagnostic metrics, which includes things like app and OS
versions, locale, and more.

Reports are queued for up to 30 days and sent when the user is online.

I tried to test as much as I could, but some of it is really tricky to
check automatically.

See also: 39773c54826ed1f13a679c835d0bfd81a24e719d