chore: use more lenient type for `createProject`
This is a types-only change.

Soon, we plan to add new properties to `api.createProject`. We'll need
to update this function parameter to support it.

Rather than change this function signature every time, let's just make
it match `api.createProject` so we don't have to keep updating it every
time.