chore: use GitHub App token to create PR in workflow
Fixes #662

This switches the github token used to create a pull request for a release candidate to use a token from a GitHub App that I have created for Awana Digital called "Awana PR Bot".

The reason for this change is so that PR checks run when the PR is opened by the [`create-release-candidate.yml`](.github/workflows/create-release-candidate.yml) workflow. Without this change, workflows are not triggered by the PR being opened.