fix: simplify "accurate location" hook (#707)
This hook handled a case where no location was provided, but that could
never happen (see 306aa518063779971fd97f417fbc95294ca975bc). This
simplifies the code because we don't need to handle that case.

I think this is a useful change on its own, but it will also be required
when we update to the latest `@comapeo/schema`.

(cherry picked from commit a8c7390e1f0e5efa34242695b631334e5f75a738)