fix: create test data with proper metadata (#708)
When creating test data, this will create a "manual location"
observation ¼ of the time. The rest of the time, it will populate fake
metadata properly.

I think this is a useful change on its own, but it will also be required
when we update to the latest `@comapeo/schema`.

(cherry picked from commit 81514f9c18bbba243625c8ec6f9d53d0b073b191)