move mmkv-related variables to constants file for now
ideally they should live in App.tsx but because of how the
useDraftObservation hook is implemented (non-pure), it relies on being
able to access a shared variable. If the variable lives in App.tsx, it
results in the corresponding test for that hook to fail because the
require graph includes the require graph of App.tsx, which is basically
the whole app.