test: speed up Jest tests by transforming fewer files
Our Jest tests can be pretty slow. After some profiling[^a], ~80% of the
time is spent transforming files with Babel. So how can we speed up
this transformation?

We can't remove transformations entirely. We need transformations for:

- TypeScript and Flow compilation (the latter is used in React Native)
- Jest module mocking
- Turning ES Modules (`import` and `export`) into CommonJS (`require`
  and `module.exports`), because Jest lacks full support for ESM
- probably other things

Jest's default transformer, `babel-jest`, seems to be the best way to do
this. Though we could get better performance with something like esbuild
or SWC[^b], those tools can't do Jest module mocking.

Before this change, though it *looked* like we were being selective,
we're actually transforming almost every file. We effectively had some
code like this (simplified slightly):

```javascript
// ...
transform: {
  '\\.[jt]sx?$': 'babel-jest',
},
transformIgnorePatterns: [
  'node_modules/(?!(...|react-native|some-others))',
],
// ...
```

See the error? We're ignoring transformation for all Node modules except
for those that start with *any three characters*...which is nearly all
of them, save a couple of two-letter modules like `ms`.
That means that we were running Babel on way more files than
necessary.[^c]

I fixed this by only transforming the modules we actually want to
transform. Unfortunately, there were a bunch of them, and it was a
simple game of whack-a-mole to find them all.

This has the advantage of speeding up tests by about 2x on my machine.
It has the disadvantage that a new/changed module could yield a
somewhat-confusing Jest error, but it *does* mention looking at
`transformIgnorePatterns` so it's not the worst. Another con: we have a
big list of modules in our Jest config.

[^a]: The simplest "profiling": clear the Jest cache, run the tests,
then run them again. They're much faster! You can also see this by
[proxying our calls to `babel-jest` and recording durations][0]; in my
testing, ~80% of test time was spent transforming with Babel.

[^b]: Both of which I tried...

[^c]: This was introduced in ecc62135945d87c44f4f38ec8e0ca962e24b726e
and is likely because [`@rnmapbox/maps` documents a code snippet][1]
that is misleading.

[0]: https://gist.github.com/EvanHahn/ac7e66460af36478235e0de17e01245c
[1]: https://github.com/rnmapbox/maps/tree/003570ff9ce3ffd6daab45c1811c5f81488122b6?tab=readme-ov-file#testing-with-jest