chore: remove useOnBackgroundedForegrounded() hook (#1304)
This hook is used to communicate with the backend when the app goes into the foreground or background. However, the backend already has an API to detect app foreground and background, which we already use. This change removes frontend code and reduces traffic over the bridge. Behaviour should remain the same.