fix: Do not stop the discovery server
Stopping the discovery server attempts to close the `node:net` server in the backend. Closing a node server awaits any active connections, and currently we never close peer connections, so if a device has any active incoming peer connections, the server will never actually close. Also, server close is non-cancellable, so there is no way to start it again. The short-term solution to this is to never try to close the discovery server.