chore: Update core native deps
This updates the backend @comapeo/core dependencies to the latest versions:

- `fs-native-extensions@1.4.3`
- `simdle-native@1.3.7`
- `quickbit-native@2.4.8`
- `sodium-native@4.3.3`

These latest versions all switch from using node-gyp-build to using cmake and bare-addon. This required adding additional patches to nodejs-mobile-react-native build.gradle because bare-addon resolves addon modules to the prebuilds folder, where-as node-gyp-build resolved addons to the build/Release folder.

The prebuilds of fs-native-extensions and quickbit-native also required patches to the build in order to build successfully for Android, but these patches seem to work at runtime from testing in an emulator.